requirements.txt
oauthlib==3.1.0
pandas==1.2.3
python-dotenv==0.15.0
requests==2.25.1
requests-oauthlib==1.3.0
SQLAlchemy==1.4.1
tweepy==3.10.0